wchar_t is two bytes and unsigned on the C6000, as TI's is
Measured with cl6x on the Windows box (scratchpad/ti/wc.c): sizeof 2,
`(wchar_t)-1 < 0` false, WCHAR_MIN 0, WCHAR_MAX 65535, L"ab" six bytes -
the Windows shape, not the Linux one this target had copied. wcharType is
UShort and lib/stddef.h's C6000 branch says `unsigned short`; a wide
literal's bytes and alignment followed. st_wide_literals measures itself
against sizeof(wchar_t) and passes as before; its tms6747 digest is the
one that moved.

+++ lib/stddef.h
@@ -27,10 +27,11 @@ typedef unsigned short     wchar_t;
 
 // ILP32 on the C6000, and TI's headers say `unsigned int`, not `unsigned
 // long` - the same width, but a different type, and in C++ a different
-// letter in every mangled name that takes a size.
-typedef unsigned int size_t;
-typedef int          ptrdiff_t;
-typedef int          wchar_t;
+// letter in every mangled name that takes a size. wchar_t is two bytes and
+// unsigned, measured with cl6x: WCHAR_MIN 0, WCHAR_MAX 65535, L"ab" six.
+typedef unsigned int   size_t;
+typedef int            ptrdiff_t;
+typedef unsigned short wchar_t;
 
 #else
